requirements.txt
elevenlabs
deepgram-sdk
openai
python-dotenv
tk
sounddevice
numpy